refactor: optics-tests/optics-deploy
