--- typescript/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-
-#Hardhat files
-cache
-artifacts
