+++ typescript/optics-deploy/.prettierrc
@@ -1,0 +1,5 @@
+{
+    "tabWidth": 2,
+    "singleQuote": true,
+    "trailingComma": "all"
+}
