+++ typescript/optics-deploy/package.json
@@ -1,0 +1,19 @@
+{
+  "devDependencies": {
+    "ethers": "^5.3.1",
+    "prettier": "^2.3.1",
+    "typescript": "^4.3.2"
+  },
+  "name": "optics-deploy",
+  "version": "1.0.0",
+  "description": "Optics deploy tools",
+  "main": "src/index.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Celo Labs Inc.",
+  "license": "MIT OR Apache-2.0",
+  "dependencies": {
+    "@ethersproject/experimental": "^5.3.0"
+  }
+}
