+++ typescript/optics-tests/package.json
@@ -7,7 +7,7 @@
     "@types/mocha": "^8.2.2",
     "chai": "^4.3.4",
     "ethereum-waffle": "^3.3.0",
-    "ethers": "^5.3.0",
+    "ethers": "^5.3.1",
     "hardhat": "^2.3.0",
     "mkdirp": "^1.0.4",
     "prettier": "2.3.0",
@@ -17,8 +17,7 @@
   },
   "dependencies": {
     "@ethersproject/experimental": "^5.3.0",
-    "@types/node": "^15.6.1",
-    "ethers": "^5.2.0"
+    "@types/node": "^15.6.1"
   },
   "name": "optics-ts",
   "version": "1.0.0",
